week5/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
blinker==1.9.0
    # via flask
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.2
    # via emotion-inference (pyproject.toml)
fsspec==2025.10.0
    # via
    #   huggingface-hub
    #   torch
gunicorn==21.2.0
    # via emotion-inference (pyproject.toml)
huggingface-hub==0.36.0
    # via
    #   tokenizers
    #   transformers
idna==3.11
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.2.6
    # via
    #   emotion-inference (pyproject.toml)
    #   opencv-python
    #   transformers
opencv-python==4.12.0.88
    # via emotion-inference (pyproject.toml)
packaging==25.0
    # via
    #   gunicorn
    #   huggingface-hub
    #   transformers
pillow==12.0.0
    # via emotion-inference (pyproject.toml)
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
regex==2025.10.23
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.6.2
    # via transformers
setuptools==80.9.0
    # via torch
sympy==1.14.0
    # via torch
tokenizers==0.22.1
    # via transformers
torch==2.9.0
    # via emotion-inference (pyproject.toml)
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.57.1
    # via emotion-inference (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
